Added binary for archlinux crosscompiler (64 bit). No beeps at startup ;) Minor other changes.

+++ include/stdint.h
@@ -16,15 +16,15 @@
 #ifndef _STDINT_H
 #define _STDINT_H
 
-typedef signed char         int8_t;
-typedef unsigned char       uint8_t;
-typedef short               int16_t;
-typedef unsigned short      uint16_t;
-typedef int                 int32_t;
-typedef unsigned            uint32_t;
+typedef signed char		int8_t;
+typedef unsigned char	uint8_t;
+typedef short			int16_t;
+typedef unsigned short	uint16_t;
+typedef int				int32_t;
+typedef unsigned		uint32_t;
 #if ARCH == 64
-typedef long long           int64_t;
-typedef unsigned long long  uint64_t;
+typedef long			int64_t;
+typedef unsigned long	uint64_t;
 #endif
 
 // TODO least, fast, greatest
